gas values optional
