+++ forc-plugins/forc-client/src/cmd/deploy.rs
@@ -1,10 +1,11 @@
 use clap::Parser;
 use fuel_crypto::SecretKey;
 
-pub use crate::util::Target;
 pub use forc::cli::shared::{BuildOutput, BuildProfile, Minify, Pkg, Print};
 pub use forc_tx::{Gas, Maturity};
 pub use forc_util::tx_utils::Salt;
+
+use crate::NodeTarget;
 
 #[derive(Debug, Default, Parser)]
 #[clap(bin_name = "forc deploy", version)]
@@ -19,6 +20,8 @@ pub struct Command {
     pub gas: Gas,
     #[clap(flatten)]
     pub maturity: Maturity,
+    #[clap(flatten)]
+    pub node: NodeTarget,
     /// Optional 256-bit hexadecimal literal(s) to redeploy contracts.
     ///
     /// For a single contract, use `--salt <SALT>`, eg.: forc deploy --salt 0x0000000000000000000000000000000000000000000000000000000000000001
@@ -38,11 +41,6 @@ pub struct Command {
     pub build_output: BuildOutput,
     #[clap(flatten)]
     pub build_profile: BuildProfile,
-    /// The URL of the Fuel node to which we're submitting the transaction.
-    /// If unspecified, checks the manifest's `network` table, then falls back
-    /// to [`crate::default::NODE_URL`].
-    #[clap(long, env = "FUEL_NODE_URL")]
-    pub node_url: Option<String>,
     /// Sign the transaction with default signer that is pre-funded by fuel-core. Useful for testing against local node.
     #[clap(long)]
     pub default_signer: bool,
@@ -54,12 +52,4 @@ pub struct Command {
     /// Sign the deployment transaction manually.
     #[clap(long)]
     pub manual_signing: bool,
-    /// Use preset configurations for deploying to a specific target.
-    ///
-    /// Possible values are: [beta-1, beta-2, beta-3, beta-4, local]
-    #[clap(long)]
-    pub target: Option<Target>,
-    /// Use preset configuration for the latest testnet.
-    #[clap(long)]
-    pub testnet: bool,
 }
