+++ forc-plugins/forc-client/src/constants.rs
@@ -1,0 +1,6 @@
+/// Default to localhost to favour the common case of testing.
+pub const NODE_URL: &str = sway_utils::constants::DEFAULT_NODE_URL;
+pub const BETA_2_ENDPOINT_URL: &str = "https://node-beta-2.fuel.network";
+pub const BETA_3_ENDPOINT_URL: &str = "https://beta-3.fuel.network";
+pub const BETA_4_ENDPOINT_URL: &str = "https://beta-4.fuel.network";
+pub const BETA_4_FAUCET_URL: &str = "https://faucet-beta-4.fuel.network";
